+++ src/components/CardRegistrationPage/NameInput.tsx
