+++ src/components/CardRegistrationPage/PasswordInput.tsx
